requirements.txt
moviepy
ffmpeg
torch
transformers
sense2vec
sentence-transformers
textwrap3
nltk
numpy
flashtext
scikit-learn
similarity
openai-whisper
git+https://github.com/boudinfl/pke.git
strsim
datasets
transformers[sentencepiece]
sentencepiece
pydub